requirements.txt
nltk==3.8.1
scikit-learn==1.3.0
pandas==2.1.0
numpy==1.24.3
pytest==7.4.0
python-dotenv==1.0.0
streamlit==1.31.0
plotly==5.18.0
pyyaml==6.0.1 